requirements.txt
torch>=2.0
transformers>=4.40
accelerate
bitsandbytes
sentencepiece
huggingface_hub 
fasttext-wheel
datasets>=2.18
evaluate
sacrebleu
sentence_splitter
spacy==3.7.2
scikit-learn
rank_bm25
bm25s
PyStemmer
numpy<2.0.0
pydantic<2.0.0
unbabel-comet
thinc==8.2.1